video_downloader/pyproject.toml
[project]
name = "video-downloader"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "yt-dlp (>=2025.12.8,<2026.0.0)",
    "confluent-kafka (>=2.12.2,<3.0.0)",
    "google-cloud (>=0.34.0,<0.35.0)",
    "google-cloud-storage (>=3.7.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.10"
